refactor(tests): move analytics test fixture into (local) conftest
We will need the same fixture in other analytics tests as well. Not
moving it to the top level conftest however, as it will most likely not
be used outside the analytics module
